pyproject.toml
[project]
name = "phantom-seed"
version = "0.1.0"
description = "A roguelike visual novel with real-time AI-generated content"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5",
    "google-genai>=1.0",
    "pydantic>=2.10",
    "pillow>=11.0",
    "rembg>=2.0.72",
    "langchain-core>=0.3",
    "langchain-google-genai>=2.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
phantom-seed = "phantom_seed.main:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"
